Remove old tests and rename new tests according to the name convention
